+++ pwa/templates/serviceworker.js
@@ -12,6 +12,16 @@ var filesToCache = [
     '/static/images/icons/icon-192x192.png',
     '/static/images/icons/icon-384x384.png',
     '/static/images/icons/icon-512x512.png',
+    '/static/images/icons/splash-640x1136.png',
+    '/static/images/icons/splash-750x1334.png',
+    '/static/images/icons/splash-1242x2208.png',
+    '/static/images/icons/splash-1125x2436.png',
+    '/static/images/icons/splash-828x1792.png',
+    '/static/images/icons/splash-1242x2688.png',
+    '/static/images/icons/splash-1536x2048.png',
+    '/static/images/icons/splash-1668x2224.png',
+    '/static/images/icons/splash-1668x2388.png',
+    '/static/images/icons/splash-2048x2732.png'
 ];
 
 // Cache on install